restauran/requirements.txt
defusedxml~=0.7.1
numpy~=2.1.0
future~=1.0.0
Pillow~=9.2.0
pip~=24.2
attrs~=24.2.0
cryptography~=43.0.3
docutils~=0.21.2
jinja2~=3.1.4
pytz~=2022.7
PyYAML~=6.0
Django~=5.1.3
asgiref~=3.8.1
sqlparse~=0.5.2
mysqlclient~=2.1.1
psycopg2~=2.9.9
colorama~=0.4.6
requests~=2.28.1
djangorestframework~=3.15.2
packaging~=23.2
inflection~=0.5.1
uritemplate~=4.1.1
setuptools~=70.0.0
pyparsing~=3.1.4
django-crispy-forms~=2.3
django-filter~=24.3
pygments~=2.18.0